requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
Authlib==1.6.1
av==15.0.0
backoff==2.2.1
beautifulsoup4==4.13.4
botocore==1.40.0
browser-use==0.5.7
bubus==1.4.7
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
constructs==10.4.2
cryptography==45.0.5
DafnyRuntimePython==4.9.0
deepdiff==8.5.0
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
e2b==1.10.0
e2b-code-interpreter==1.5.2
email_validator==2.2.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.28.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.30.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsii==1.113.0
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
limits==5.4.0
livekit==1.0.12
livekit-agents==1.2.2
livekit-api==1.0.5
livekit-blingfire==1.0.0
livekit-plugins-cartesia==1.2.2
livekit-plugins-deepgram==1.2.2
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-openai==1.2.2
livekit-plugins-silero==1.2.2
livekit-plugins-turn-detector==1.2.2
livekit-protocol==1.0.4
logfire==4.0.1
markdown-it-py==3.0.0
markdown-pdf==1.5
markdownify==1.1.0
MarkupSafe==3.0.2
mcp==1.12.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
nest-asyncio==1.6.0
numpy==2.3.2
oauthlib==3.3.1
ollama==0.5.1
onnxruntime==1.22.1
openai==1.98.0
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orderly-set==5.5.0
packaging==25.0
patchright==1.52.5
pathable==0.4.4
pgvector==0.4.1
pillow==11.3.0
playwright==1.54.0
portalocker==2.10.1
posthog==6.3.2
prometheus_client==0.22.1
propcache==0.3.2
propelauth-fastapi==4.2.8
propelauth-py==4.2.8
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
publication==0.0.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.25.3
pyparsing==3.2.3
pypdf==5.9.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.5.3
screeninfo==0.8.1
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SQLAlchemy==2.0.42
sse-starlette==3.0.2
starlette==0.47.2
stripe==12.4.0
svix==1.70.0
sympy==1.14.0
tenacity==8.5.0
tokenizers==0.21.4
tqdm==4.67.1
transformers==4.54.1
typeguard==2.13.3
typer==0.16.0
types-Deprecated==1.2.15.20250304
types-protobuf==4.25.0.20240417
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uuid7==0.1.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
